fix: Handle XGetClassHint() failures
`XGetClassHint()` will not initialize passed in `hint` structure, if the
call fails.  Accessing `hint.res_name` and `hint.res_class` when
`XGetClassHint()` failed accesses random memory.

Also, refactored `add_icon` error handing a bit.  Hopefully, it is
easier to read in this form.  As there are a few cases when `classname`
would be unnecessary, I've moved `classname` initialization to be after
those are checked.